+++ NEWS.md
@@ -5,7 +5,7 @@ This is a leader-intensive update. Additions/new features include:
 
 - `LEAD`: a data set on select leader experience and attribute descriptions (LEAD).
 - `lwuf`: a data set from [Carter and Smith (2020)](https://doi.org/10.1017/S0003055420000313) measuring leader willingness to use force.
-- `download_extdata()` now also downloads directed leader dyad-year data for all directed leader dyad-years from 1870 to 2015.
+- `download_extdata()` now also downloads directed leader dyad-year data for all directed leader dyad-years from 1870 to 2015. The data come in two forms. One is all possible (inter-state, leader-overlap) dyadic leader pairings from 1870 to 2015, as presented in the Archigos data. The second standardizes these data to the CoW state system data.
 - `create_leaderdyadyears()`: create leader dyad-year data (via `archigos`) for all leader dyad-year combinations from 1870 to 2015. This function leans on remote data available for download in `download_extdata()`.
 - `gml_part`: a data set on participant summary information from the GML MID data.
 - `add_gml_mids()` now works with leader-year data. See documentation for more.